+++ exspy/_docstrings/model.py
